Partition size - use df instead of df -h
Partition size tests now use the 1k blocks from df, instead of using
df -h. Kudos to Marcus Chang for spotting this. The test was pulling
directly the float value from the df -h result and this yields interesting
results with larger memory cards, as some of partitions are scaled in size
and when comparing floats 1G is less than 700MB because the float conversion
ignored the types completely. By using the direct 1K blocks instead we
avoid this problem and the tests become less brittle.